perf(export class): export component class #195
This commit updated the button component to export the class
by default and supports extension of the core element. This update
also addresses bundling performance updated.

Changes to be committed:
modified:   .github/workflows/testPublish.yml
modified:   demo/index.html
modified:   index.js
modified:   package-lock.json
modified:   package.json
deleted:    rollup.config.js
new file:   rollup.config.mjs
modified:   src/auro-button.js
