+++ demo/index.html
@@ -27,7 +27,7 @@
           Prism.highlightAll();
         });
     </script>
-    <script type="module" src="../src/auro-button.js"></script>
+    <script type="module" src="./../index.js"></script>
     <script src="https://unpkg.com/@alaskaairux/auro-accordion@latest/dist/auro-accordion__bundled.js" type="module"></script>
     <script src="https://unpkg.com/@alaskaairux/auro-icon@latest/dist/auro-icon__bundled.js" type="module"></script>
     <script src="https://unpkg.com/@alaskaairux/auro-alerts@latest/dist/auro-alerts__bundled.js" type="module"></script>
