+++ package.json
@@ -38,26 +38,17 @@
     "@alaskaairux/design-tokens": "^3.12.2",
     "@alaskaairux/icons": "^4.22.0",
     "@alaskaairux/webcorestylesheets": "^3.7.3",
-    "@babel/core": "^7.15.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.15.0",
-    "@babel/preset-env": "^7.15.0",
-    "@babel/runtime": "^7.15.3",
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-conventional": "^11.0.0",
     "@open-wc/testing": "^2.5.33",
     "@open-wc/testing-karma": "^4.0.9",
-    "@rollup/plugin-alias": "^3.1.5",
-    "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-node-resolve": "^11.2.1",
+    "@rollup/plugin-node-resolve": "^15.0.1",
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.0",
     "@semantic-release/npm": "^7.1.3",
     "@web/dev-server": "^0.1.21",
     "@webcomponents/webcomponentsjs": "^2.6.0",
     "autoprefixer": "^10.3.1",
-    "babel-loader": "^8.2.2",
     "compression": "^1.7.4",
     "concat": "^1.0.3",
     "copyfiles": "^2.4.1",
@@ -67,7 +58,6 @@
     "focus-visible": "^5.2.0",
     "husky": "^4.3.8",
     "lodash": "^4.17.21",
-    "marked": "^4.0.10",
     "nodemon": "^2.0.21",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.3.6",
@@ -76,10 +66,8 @@
     "postcss-remove-rules": "^1.0.0",
     "postcss-selector-replace": "^1.0.2",
     "prismjs": "^1.24.1",
-    "rollup": "^2.56.2",
-    "rollup-plugin-minify-html-literals": "^1.2.6",
-    "rollup-plugin-serve": "^1.1.0",
-    "rollup-plugin-terser": "^7.0.2",
+    "rollup": "^3.19.1",
+    "rollup-plugin-serve": "^2.0.2",
     "semantic-release": "^19.0.3",
     "sinon": "^9.2.4",
     "stylelint": "^13.13.1",
