+++ rollup.config.mjs
@@ -1,0 +1,24 @@
+import { nodeResolve } from '@rollup/plugin-node-resolve';
+import serve from 'rollup-plugin-serve';
+
+const production = !process.env.ROLLUP_WATCH;
+
+const modernConfig = {
+  input: {
+    ['auro-badge__bundled']: './index.js',
+  },
+  output: {
+    format: 'esm',
+    dir: 'dist/'
+  },
+  plugins: [
+    nodeResolve(),
+    !production &&
+      serve({
+        open: true,
+        openPage: '/docs/'
+      })
+  ]
+};
+
+export default [modernConfig];
